[RF] Improve documentation of `RooAbsData::split()`
In #13087, it was noted that `RooAbsData::split()` does not split
observables. However, there is already one overload of this function to
do this, taking a RooSimultaneous as an input to know what the
observables are. This PR improves the docs of `RooAbsData::split()` by
adding a comment that links to this other overload, such that it is
easier to find.

Closes #13087.