[RF] Fix unbinned model in `test_lib.h`
The unbinned model in that was used by the `testLikelihoodSerial` and
the `estLikelihoodJob` unit tests was corrupt, which resulted in
undefined behavior.

So far, it was the RooAddPdf of `2 * n` pdfs, with `2 * n` yield
coefficients. One component for sinal, and one for background in each of
the `n` dimensions. However, this makes no sense for many reasons. Just
to give an example, the backround and the signal sample should only have
one coefficient each, since they have the same yield in each of the
dimensions (the different observables are just different observables for
*the same event*, other than in the `RooSimultaneous` case where we have
different observables for *independent events*).

In RooFit, pdfs that are for different obserables of of the same event
should be stiched together with the `RooProdPdf`.

Now, it was redefined to be the `RooAddPdf` of two n-dimensional
RooProdPdfs for signal and background, which is a very common RooFit
pattern.