[RF] Rename RooIntegrator1D to RooRombergIntegrator and add unit test
Now that the RooIntegrator1D class can handle the n-dimensional case, it
should be renamed. It would be appropriate to put "Romberg" in the name,
because the class is implementing Rombergs method.

Also, a new unit test that checks the results of the algorithms it
provides are the same as in ROOT 6.28.04 is added.