[RF][HS3][HF] Fix problems with `ShapeSys` in HistFactory HS3
The JSON export of HistFactory models did way too much guess work when
figuring out the name of the `ShapeSys` names. In particular, it tried
to to remove a prefix with the sample name, which is the wrong reverse
engineering of what HistFactory does and broke workspaces where the user
prefixed the `ShapeSys` names with the sample names.