[RF] Test cleanup after exception in an evt loop
Add a unit test for the cleanup happening after the event loop is
interrupted by an exception.
Moreover, adapt two existing unit tests to the new implementation of
RLoopManager::Run. The two tests relied on a faulty behaviour of RDF,
which allowed to trigger subsequent event loops by dereferencing
RResultPtr's instances and trapping the exception thrown during the
event loop in a try-catch block.