[cmake,foundation] Do not generate `RGitCommit.h` but `etc/gitinfo.txt`:
Deprecate the use of `RGitCommit.h`: it triggers rebuilds where the git info
is in fact not needed at compiletime; all uses are replaced by parsing
`gitinfo.txt`.

Instead, create `etc/gitinfo.txt` even if the source directory is not a git
repo. This only works for releases (even `ROOT_PATCH_VERSION`), where the
tag and the branch can be determined from the version numbers.