[RF] Loosen comparision precision in `testHypoTestInvResult`
For the estimated error on the upper limit, we need to loosen the
tolerance because it is determined on-the-fly with a fit, and not read
as a data member like the comment in the test said before.

Withg the default minimizer changing from Minuit to Minuit2, the results
are slightly different and caused this test to fail.