[RF][HS3][HF] Backwards compatible way of dealing with bin widths
The bin width modifiers are handled separately from the other modifiers
in the HistFactory JSON IO. However, do detect them, we can't just check
for the RooBinWidthFunction type here, because prior to ROOT 6.26, the
multiplication with the inverse bin width was done in a different way
(like a normfactor with a RooRealVar, but it was stored in the dataset).
Fortunately, the name was similar, so we can match the modifier name.