[RF] Also support `Offset("bin")` in simul. fits with legacy backend
Also support the `Offset("bin")` feature in simultaneous fits with the
legacy evaluation backend (e.g., the old test statistic classes).

In this case, The top-level `RooNLLVar` actually doesn't do anything
but contains pointers to the NLLs for the individual channels. The
function call to enable the offsetting must be delegated to those.

A unit test that covers this is also implemented.

Note that for the new CPU evaluation backend, this is working already.