[RF] Migrate `stressHistFactory` to gtest in `histfactory/test`
It was not meaningful to have a clone on the `stressRooFit` tests suite
just for a single test that compares pdf values, and it didn't even use
the reference file mechanism from `RooUnitTest`.

Since future unit tests are not added to the `stress` suites anymore but
added as gtests, we might as well also migrate the single test that was
formerly `stressHistFactory`.