[cmake] Look for `CUDAToolkit` when building with `tmva-gpu`
This is a follow-up to 15a14ae6b4c.

The `FindCUDA` module is deprecated since CMake 3.10 and the mentioned
commit therefore removed the usage of it.

This was fine for RooFit, but TMVA not only compiles sources with the
NVidia compiler but also uses the CUDA toolkit in regular C++.

Therefore, that commit broke the TMVA GPU features.

Unfortunately, I didn't notice before, because of this check
in `cmake/modules/SearchInstalledSoftware.cmake`:
```cmake
  if(tmva-gpu AND NOT CUDA_FOUND)
    set(tmva-gpu OFF CACHE BOOL "Disabled because cuda not found" FORCE)
  endif()
```

The `CUDA_FOUND` variable is not set anymore now that `FindCUDA` is not
used. That means that `tmva-gpu` was silently disabled, and not building
successfully (which I had wrongly concluded because of the absence of
errors). This check needs to be replaced with
`if( NOT CMAKE_CUDA_COMPILER)`.

As recommended in the CMake docs, this should be done with the
`FindCUDAToolkit` module, which is what this PR suggests.

The `CUDA_ADD_EXECUTABLE` command is also replaced with the regular
`add_executable`, because `CUDA_ADD_EXECUTABLE` was also deprecated with
`FindCUDA`.

Furthermore, all `cuda AND tmva-gpu` checks were replaced with only
`tmva-gpu`, because `tmva-gpu` implies `cuda`.

See also:

  * https://cmake.org/cmake/help/latest/module/FindCUDA.html
  * https://cmake.org/cmake/help/latest/module/FindCUDAToolkit.html