Minuit2: Fix VERSION_FILE path for standalone mode
Fixes #14157. When building minuit2 standalone with root 6.30/02
(-Dminuit2_standalone=ON), cmake errors out when looking for the
version file 'RVersion.hxx' in 'core/foundation/inc/ROOT/RVersion.hxx'
whereas it is copied to the top-level dir (i.e. 'math/minuit2/'). This
is because the variable VERSION_FILE is incorrectly set to the file path
matching the full path of the file. Fixed by setting the parent dir path
of 'RVersion.hxx' as the `SOURCE` in `copy_standalone()` so that
`VERSION_FILE` is set to just the base file name.