[RF] Introduce `RooAbsCollection::addTyped()` method
It's a very common pattern in the RooFit implementation and user code to
check the type when adding to a `RooAbsCollection`, and erroring out if
the type is not correct.

This commit suggests a function in `RooAbsCollection` to do exactly
this, drastically reducing the code duplication in RooFit and experiment
frameworks like CMS combine.